requirements.txt
aniso8601==9.0.1
attrs==23.1.0
click==8.0.3
Flask==3.0.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.17.3
MarkupSafe==2.1.1
marshmallow==3.20.1
PyJWT=2.3.0
pyrsistent==0.20.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
Werkzeug==3.0.1
zipp==3.6.0
